run tests against multiple kestra versions (#206)
* try go

* ci

* ci

* refactor ci

* fix path

* fix s

* d

* Update .github/workflows/java-sdk.yml


* fix

---------

+++ python/python-sdk/run-tests.sh
@@ -8,26 +8,13 @@ log_and_run() {
 }
 
 if [ $# -ge 1 ]; then
-  KESTRA_VERSION="$1"
+  versions="$1"
 else
-  KESTRA_VERSION=$(cat ../../COMPATIBLE_KESTRA_VERSION.properties)
+  versions=$(cat ../../COMPATIBLE_KESTRA_VERSION.properties)
 fi
 
-export KESTRA_VERSION=$KESTRA_VERSION
-
 echo "/n------------------------------------------------"
-echo "Build local SDK and test it in an example project"
-echo "docker KESTRA_VERSION used: $KESTRA_VERSION\n"
-
-
-echo "start Kestra container"
-log_and_run docker compose -f docker-compose-ci.yml down
-
-log_and_run docker compose -f docker-compose-ci.yml up -d --wait || {
-   echo "db Docker Compose failed. Dumping logs:";
-   log_and_run docker compose -f docker-compose-ci.yml logs;
-   exit 1;
-}
+echo "Build local SDK and test it in a docker Kestra instance"
 
 echo "install requirements"
 log_and_run pip install -r requirements.txt -r test-requirements.txt --break-system-packages
@@ -35,8 +22,27 @@ log_and_run pip install -r requirements.txt -r test-requirements.txt --break-sys
 echo "install SDK locally so it can be imported and used in e2e tests"
 log_and_run pip install -e . --break-system-packages
 
-echo "start tests"
-log_and_run python3 -m pytest -vv -s --log-cli-format="%(asctime)s [%(levelname)s] %(name)s: %(message)s)" --showlocals --timeout=10
+for KESTRA_VERSION in $versions; do
+  if [ -z "$KESTRA_VERSION" ]; then
+    continue
+  fi
 
-echo "stop Kestra container"
-log_and_run docker compose -f docker-compose-ci.yml down
+  echo "docker KESTRA_VERSION used: $KESTRA_VERSION\n"
+
+  export KESTRA_VERSION=$KESTRA_VERSION
+
+  echo "start Kestra container"
+  log_and_run docker compose -f docker-compose-ci.yml down
+
+  log_and_run docker compose -f docker-compose-ci.yml up -d --wait || {
+     echo "db Docker Compose failed. Dumping logs:";
+     log_and_run docker compose -f docker-compose-ci.yml logs;
+     exit 1;
+  }
+
+  echo "start tests"
+  log_and_run python3 -m pytest -vv -s --log-cli-format="%(asctime)s [%(levelname)s] %(name)s: %(message)s)" --showlocals --timeout=10
+
+  echo "stop Kestra container"
+  log_and_run docker compose -f docker-compose-ci.yml down
+done
